split each into its own component
